Switch from slog to tracing.
The custom tracing subscriber is hacky and a bit unsafe, in basically
the same ways as the slog implementation was. The tracing API doesn't
give us a way to guarantee the type of the value being logged is what
we expect it to be, so we just have to assume it's correct.